feat: add sampling rules GraphQL API and RBAC permissions
Add CRUD operations for sampling rules via GraphQL, refactor config
services, and grant samplings resource access in UI RBAC roles.

Made-with: Cursor
