+++ frontend/services/odigos_config.go
@@ -2,16 +2,16 @@ package services
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 
+	"github.com/odigos-io/odigos/api/k8sconsts"
 	"github.com/odigos-io/odigos/common"
 	"github.com/odigos-io/odigos/common/consts"
-	"github.com/odigos-io/odigos/config"
-	"github.com/odigos-io/odigos/frontend/graph/model"
 	"github.com/odigos-io/odigos/k8sutils/pkg/env"
 
 	v1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/util/retry"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -41,8 +41,7 @@ func getOdigosConfigFromConfigMap(ctx context.Context, c client.Client, configMa
 
 // GetEffectiveConfig retrieves the current effective configuration from the effective-config ConfigMap.
 func GetEffectiveConfig(ctx context.Context, c client.Client) (*common.OdigosConfiguration, error) {
-	config, _, err := GetEffectiveConfigWithRawYAML(ctx, c)
-	return config, err
+	return getOdigosConfigFromConfigMap(ctx, c, consts.OdigosEffectiveConfigName)
 }
 
 // GetHelmDeploymentConfig retrieves the current helm deployment configuration from the odigos-helm-deployment-config ConfigMap.
@@ -60,89 +59,62 @@ func GetLocalUIConfig(ctx context.Context, c client.Client) (*common.OdigosConfi
 	return getOdigosConfigFromConfigMap(ctx, c, consts.OdigosLocalUiConfigName)
 }
 
-func GetConfigYamls() ([]*model.ConfigYaml, error) {
-	var resp []*model.ConfigYaml
-
-	for _, cfg := range config.Get() {
-		var fields []*model.ConfigYamlField
-		for _, f := range cfg.Spec.Fields {
-			field := &model.ConfigYamlField{
-				DisplayName:   f.DisplayName,
-				ComponentType: model.FieldType(f.ComponentType),
-				IsHelmOnly:    f.IsHelmOnly,
-				Description:   f.Description,
-				HelmValuePath: f.HelmValuePath,
-			}
-
-			if f.DocsLink != "" {
-				field.DocsLink = &f.DocsLink
-			}
-
-			if len(f.ComponentProps) > 0 {
-				propsJSON, err := json.Marshal(f.ComponentProps)
-				if err != nil {
-					return nil, fmt.Errorf("failed to marshal component props: %w", err)
-				}
-				s := string(propsJSON)
-				field.ComponentProps = &s
-			}
-
-			fields = append(fields, field)
-		}
-
-		resp = append(resp, &model.ConfigYaml{
-			Name:        cfg.Metadata.Name,
-			DisplayName: cfg.Metadata.DisplayName,
-			Fields:      fields,
-		})
-	}
-
-	return resp, nil
-}
-
-// GetEffectiveConfigWithRawYAML retrieves the effective config along with its raw YAML representation.
-func GetEffectiveConfigWithRawYAML(ctx context.Context, c client.Client) (*common.OdigosConfiguration, string, error) {
-	ns := env.GetCurrentNamespace()
-
-	var cm v1.ConfigMap
-	err := c.Get(ctx, types.NamespacedName{Namespace: ns, Name: consts.OdigosEffectiveConfigName}, &cm)
-	if err != nil {
-		return nil, "", client.IgnoreNotFound(err)
-	}
-
-	if cm.Data == nil || cm.Data[consts.OdigosConfigurationFileName] == "" {
-		return nil, "", nil
-	}
-
-	rawYAML := cm.Data[consts.OdigosConfigurationFileName]
-
-	var odigosConfig common.OdigosConfiguration
-	if err := yaml.Unmarshal([]byte(rawYAML), &odigosConfig); err != nil {
-		return nil, "", fmt.Errorf("failed to parse odigos config: %w", err)
-	}
-
-	return &odigosConfig, rawYAML, nil
-}
-
-func PersistUiLocalSamplingConfig(ctx context.Context, c client.Client, samplingConfig *common.SamplingConfiguration) error {
+// upsertLocalUiConfig applies a mutation to the odigos-local-ui-config ConfigMap,
+// creating it with proper owner references if it does not yet exist.
+func upsertLocalUiConfig(ctx context.Context, c client.Client, mutate func(cfg *common.OdigosConfiguration)) error {
 	ns := env.GetCurrentNamespace()
 
 	return retry.RetryOnConflict(retry.DefaultBackoff, func() error {
 		var cm v1.ConfigMap
-		err := c.Get(ctx, types.NamespacedName{Namespace: ns, Name: consts.OdigosLocalUiConfigName}, &cm)
+		if err := c.Get(ctx, types.NamespacedName{Namespace: ns, Name: consts.OdigosLocalUiConfigName}, &cm); err != nil {
+			if apierrors.IsNotFound(err) {
+				ownerCm := v1.ConfigMap{}
+				if err := c.Get(ctx, types.NamespacedName{Namespace: ns, Name: consts.OdigosConfigurationName}, &ownerCm); err != nil {
+					return fmt.Errorf("failed to get odigos-configuration for owner reference: %w", err)
+				}
+				cfg := common.OdigosConfiguration{}
+				mutate(&cfg)
+				data, marshalErr := yaml.Marshal(cfg)
+				if marshalErr != nil {
+					return marshalErr
+				}
+				newCm := v1.ConfigMap{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      consts.OdigosLocalUiConfigName,
+						Namespace: ns,
+						Labels:    map[string]string{k8sconsts.OdigosSystemConfigLabelKey: "local-ui"},
+						OwnerReferences: []metav1.OwnerReference{{
+							APIVersion: "v1", Kind: "ConfigMap", Name: ownerCm.Name, UID: ownerCm.UID,
+						}},
+					},
+					Data: map[string]string{consts.OdigosConfigurationFileName: string(data)},
+				}
+				return c.Create(ctx, &newCm)
+			}
+			return err
+		}
+
+		var cfg common.OdigosConfiguration
+		if cm.Data != nil && cm.Data[consts.OdigosConfigurationFileName] != "" {
+			if err := yaml.Unmarshal([]byte(cm.Data[consts.OdigosConfigurationFileName]), &cfg); err != nil {
+				return fmt.Errorf("parse existing config: %w", err)
+			}
+		}
+		mutate(&cfg)
+		data, err := yaml.Marshal(cfg)
 		if err != nil {
 			return err
 		}
-		config := common.OdigosConfiguration{}
-		if err := yaml.Unmarshal([]byte(cm.Data[consts.OdigosConfigurationFileName]), &config); err != nil {
-			return err
+		if cm.Data == nil {
+			cm.Data = make(map[string]string)
 		}
-		config.Sampling = samplingConfig
-		yamlText, err := yaml.Marshal(config)
-		if err != nil {
-			return err
-		}
-		cm.Data[consts.OdigosConfigurationFileName] = string(yamlText)
+		cm.Data[consts.OdigosConfigurationFileName] = string(data)
 		return c.Update(ctx, &cm)
+	})
+}
+
+func PersistUiLocalSamplingConfig(ctx context.Context, c client.Client, samplingConfig *common.SamplingConfiguration) error {
+	return upsertLocalUiConfig(ctx, c, func(cfg *common.OdigosConfiguration) {
+		cfg.Sampling = samplingConfig
 	})
 }
