+++ helm/odigos/templates/ui/role.yaml
@@ -60,6 +60,7 @@ rules:
       - 'instrumentationrules'
       - 'destinations'
       - 'actions'
+      - 'samplings'
     verbs:
       - create
       - patch
